# online-book-store/src/main/java/ae/smartdubai/bookstore/model/CheckoutResponse.java
package ae.smartdubai.bookstore.model;

import java.math.BigDecimal;

public class CheckoutResponse {

	private BigDecimal totalPrice;

	public BigDecimal getTotalPrice() {
		return totalPrice;
	}

	public void setTotalPrice(BigDecimal totalPrice) {
		this.totalPrice = totalPrice;
	}
}